requirements.txt
insightface==0.7.3
opencv-python==4.11.0.86
psycopg2-binary==2.9.10
bcrypt==4.3.0
numpy==1.23.5


